add to Job Type Api delete and update services

+++ src/main/java/com/umutdoruk/hrms/repository/JobTypeRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobTypeRepository extends JpaRepository<JobType, Integer> {
-
+public interface JobTypeRepository extends JpaRepository<JobType, Long> {
 
 }
